as.integer64.character {bit64}	R Documentation

Coerce to integer64

Description

Methods to coerce from other atomic types to integer64.

Usage

 NA_integer64_
 as.integer64(x, ...)
 ## S3 method for class 'integer64'
as.integer64(x, ...)
 ## S3 method for class 'NULL'
as.integer64(x, ...)
 ## S3 method for class 'character'
as.integer64(x, ...)
 ## S3 method for class 'bitstring'
as.integer64(x, ...)
 ## S3 method for class 'double'
as.integer64(x, keep.names = FALSE, ...)
 ## S3 method for class 'integer'
as.integer64(x, ...)
 ## S3 method for class 'logical'
as.integer64(x, ...)
 ## S3 method for class 'factor'
as.integer64(x, ...)

Arguments

x
an atomic vector

keep.names
FALSE, set to TRUE to keep a names vector

...
further arguments to the NextMethod

Details

as.integer64.character is realized using C function strtoll which does not support scientific notation. Instead of '1e6' use '1000000'. as.integer64.bitstring evaluates characters '0' anbd ' ' as zero-bit, all other one byte characters as one-bit, multi-byte characters are not allowed, strings shorter than 64 characters are treated as if they were left-padded with '0', strings longer than 64 bytes are mapped to NA_INTEGER64 and a warning is emitted.

Value

The other methods return atomic vectors of the expected types

Author(s)

Jens Oehlschlägel <Jens.Oehlschlaegel@truecluster.com>

See Also

as.character.integer64 integer64

Examples

as.integer64(as.character(lim.integer64()))
as.integer64(
  structure(c("1111111111111111111111111111111111111111111111111111111111111110", 
              "1111111111111111111111111111111111111111111111111111111111111111", 
              "1000000000000000000000000000000000000000000000000000000000000000",
              "0000000000000000000000000000000000000000000000000000000000000000", 
              "0000000000000000000000000000000000000000000000000000000000000001", 
              "0000000000000000000000000000000000000000000000000000000000000010" 
  ), class = "bitstring")
)
as.integer64(
 structure(c("............................................................... ", 
             "................................................................", 
             ".                                                               ",
             "", 
             ".", 
             "10"
  ), class = "bitstring")
)

[Package bit64 version 4.0.5 Index]