duplicated.integer64 {bit64}	R Documentation

Determine Duplicate Elements of integer64

Description

duplicated() determines which elements of a vector or data frame are duplicates of elements with smaller subscripts, and returns a logical vector indicating which elements (rows) are duplicates.

Usage

## S3 method for class 'integer64'
duplicated(x, incomparables = FALSE, nunique = NULL, method = NULL, ...)

Arguments

x
a vector or a data frame or an array or NULL.

incomparables
ignored

nunique
NULL or the number of unique values (including NA). Providing nunique can speed-up matching when x has no cache. Note that a wrong nunique can cause undefined behaviour up to a crash.

method
NULL for automatic method selection or a suitable low-level method, see details

...
ignored

Details

This function automatically chooses from several low-level functions considering the size of x and the availability of a cache.

Suitable methods are hashdup (hashing), sortorderdup (fast ordering) and orderdup (memory saving ordering).

Value

duplicated(): a logical vector of the same length as x.

Author(s)

Jens Oehlschlägel <Jens.Oehlschlaegel@truecluster.com>

See Also

duplicated, unique.integer64

Examples

x <- as.integer64(sample(c(rep(NA, 9), 1:9), 32, TRUE))
duplicated(x)

stopifnot(identical(duplicated(x),  duplicated(as.integer(x))))

[Package bit64 version 4.0.5 Index]