identical.integer64 {bit64}	R Documentation

Identity function for class 'integer64'

Description

This will discover any deviation between objects containing integer64 vectors.

Usage

 identical.integer64(x, y, num.eq = FALSE, single.NA = FALSE
, attrib.as.set = TRUE, ignore.bytecode = TRUE)

Arguments

x
atomic vector of class 'integer64'

y
atomic vector of class 'integer64'

num.eq
see identical

single.NA
see identical

attrib.as.set
see identical

ignore.bytecode
see identical

Details

This is simply a wrapper to identical with default arguments num.eq = FALSE, single.NA = FALSE.

Value

A single logical value, TRUE or FALSE, never NA and never anything other than a single value.

Author(s)

Jens Oehlschlägel <Jens.Oehlschlaegel@truecluster.com>

See Also

==.integer64 identical integer64

Examples

  i64 <- as.double(NA); class(i64) <- "integer64"
  identical(i64-1, i64+1)
  identical.integer64(i64-1, i64+1)

[Package bit64 version 4.0.5 Index]